Fixed timezone issue in month/year heading on calendar
The month and year strings on the monthly calendar view after a
selectDate call are generated using a date formatter that expects local
dates, but the GMT date was passed in.  In my case, calling selectDate
with an early morning time on the first day of a month was causing the
month view to print the previous month's name (e.g., "June" instead of
"July").